feat: refactor imports to improve module organization and clarity
