+++ src/index.ts
@@ -1,8 +1,5 @@
 import "./styles/index.scss";
 
-export * from "@/components";
-
 export { useDevPanel } from "@/hooks/useDevPanel";
-export { useDevPanelTheme } from "@/store/ThemeStore";
 export type { DevPanelProps } from "@/components/DevPanel/types";
 export type { ControlsGroup } from "@/components/ControlRenderer/controls/types";
